image_checker/ai-image-detector/requirements.txt
numpy
pillow
tensorflow
scikit-learn
opencv-python
matplotlib
pytest